sec(oracle): critical patch applied to crates/ast-engine/context.rs

+++ crates/ast-engine/context.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub struct PaymentContext {
-    pub amount: f64,
+    pub amount: u64, // Consider a dedicated decimal type or integer for smallest unit
     pub country: String,
     pub asset: String,
     pub risk_score: f32,
